Associate the auth and article operate(create update delete)

+++ internal/service/article.go
@@ -1,12 +1,16 @@
 package service
 
-import "Goose/internal/model"
+import (
+	"Goose/global"
+	"Goose/internal/model"
+	"Goose/pkg/errorCode"
+)
 
 func (svc *Service) CountArticle(title string, state uint8) (int, error) {
 	return svc.dao.CountArticle(title, state)
 }
 
-func (svc *Service) GetArticle(title string, createdBy string, state uint8) (*model.Article, error){
+func (svc *Service) GetArticle(title string, createdBy string, state uint8) (*model.Article, error) {
 	return svc.dao.GetArticle(title, createdBy, state)
 }
 
@@ -22,6 +26,22 @@ func (svc *Service) UpdateArticle(ID uint32, title, desc, content, coverImageUrl
 	return svc.dao.UpdateArticle(ID, title, desc, content, coverImageUrl, modifiedBy, state)
 }
 
-func (svc *Service) DeleteArticle(ID uint32) error {
-	return svc.dao.DeleteArticle(ID)
+func (svc *Service) DeleteArticle(ID uint32, authName string) *errorCode.Error {
+	// 检验删除权限
+	article, err := svc.dao.GetArticleByID(ID)
+	if err != nil {
+		global.Logger.ErrorF("svc.dao.GetArticleByID err: %v", err)
+		return errorCode.ErrorDeleteArticleFail
+	}
+	if article.CreatedBy != authName {
+		return errorCode.ErrorDeleteArticlePower
+	}
+
+	// 删除文章
+	err = svc.dao.DeleteArticle(ID)
+	if err != nil {
+		global.Logger.ErrorF("svc.DeleteArticle err: %v", err)
+		return errorCode.ErrorDeleteArticleFail
+	}
+	return nil
 }
